Crate pallet_crowdloan

Crowdloan Pallet

A pallet allowing users to create generic crowdloans and contribute to them, the raised funds are then transferred to a target address and an extrinsic is dispatched, making it reusable for any crowdloan type.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_crowdloan

Structs

CrowdloanInfo

A struct containing the information about a crowdloan.

Type Aliases

BalanceOf

BoundedCallOf

CrowdloanId

CrowdloanInfoOf

CurrencyOf