pallet_crowdloan::pallet

Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::RuntimeCall>;
    type Currency: Balanced<Self::AccountId, Balance = u64> + Mutate<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage;
    type PalletId: Get<PalletId>;
    type MinimumDeposit: Get<BalanceOf<Self>>;
    type AbsoluteMinimumContribution: Get<BalanceOf<Self>>;
    type MinimumBlockDuration: Get<BlockNumberFor<Self>>;
    type MaximumBlockDuration: Get<BlockNumberFor<Self>>;
    type RefundContributorsLimit: Get<u32>;
    type MaxContributors: Get<u32>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::RuntimeCall>

The overarching call type.

type Currency: Balanced<Self::AccountId, Balance = u64> + Mutate<Self::AccountId>

The currency mechanism.

type WeightInfo: WeightInfo

The weight information for the pallet.

type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage

The preimage provider which will be used to store the call to dispatch.

type PalletId: Get<PalletId>

The pallet id that will be used to derive crowdloan account ids.

type MinimumDeposit: Get<BalanceOf<Self>>

The minimum deposit required to create a crowdloan.

type AbsoluteMinimumContribution: Get<BalanceOf<Self>>

The absolute minimum contribution required to contribute to a crowdloan.

type MinimumBlockDuration: Get<BlockNumberFor<Self>>

The minimum block duration for a crowdloan.

type MaximumBlockDuration: Get<BlockNumberFor<Self>>

The maximum block duration for a crowdloan.

type RefundContributorsLimit: Get<u32>

The maximum number of contributors that can be refunded in a single refund.

type MaxContributors: Get<u32>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.