pallet_crowdloan

Type Alias CrowdloanInfoOf

pub type CrowdloanInfoOf<T> = CrowdloanInfo<<T as Config>::AccountId, BalanceOf<T>, BlockNumberFor<T>, BoundedCallOf<T>>;

Aliased Type

struct CrowdloanInfoOf<T> {
    pub creator: <T as Config>::AccountId,
    pub deposit: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    pub min_contribution: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    pub end: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
    pub cap: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    pub funds_account: <T as Config>::AccountId,
    pub raised: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    pub target_address: Option<<T as Config>::AccountId>,
    pub call: Option<Bounded<<T as Config>::RuntimeCall, <T as Config>::Hashing>>,
    pub finalized: bool,
    pub contributors_count: u32,
}

Fields
creator: <T as Config>::AccountId

The creator of the crowdloan.

deposit: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

The initial deposit of the crowdloan from the creator.

min_contribution: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

Minimum contribution to the crowdloan.

end: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

The end block of the crowdloan.

cap: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

The cap to raise.

funds_account: <T as Config>::AccountId

The account holding the funds for this crowdloan. Derived on chain but put here for ease of use.

raised: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

The amount raised so far.

target_address: Option<<T as Config>::AccountId>

The optional target address to transfer the raised funds to, if not provided, it means the funds will be transferred from on chain logic inside the provided call to dispatch.

call: Option<Bounded<<T as Config>::RuntimeCall, <T as Config>::Hashing>>

The optional call to dispatch when the crowdloan is finalized.

finalized: bool

Whether the crowdloan has been finalized.

contributors_count: u32

The number of contributors to the crowdloan.